third_party/WebKit/LayoutTests/css-parser/color3_hsl-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Color test for resources/css-parsing-tests/color3_hsl.json

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS colorTest.parseColor("hsl(0, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 0%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 12.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 25%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 37.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 50%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 62.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 75%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 87.5%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(30, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(60, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(90, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(120, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(150, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(180, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(210, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(240, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(270, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(300, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(330, 100%, 0%)") is "rgb(0, 0, 0)"
PASS colorTest.parseColor("hsl(0, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(30, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(60, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(90, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(120, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(150, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(180, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(210, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(240, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(270, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(300, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(330, 0%, 12.5%)") is "rgb(31, 31, 31)"
PASS colorTest.parseColor("hsl(0, 12.5%, 12.5%)") is "rgb(35, 27, 27)"
PASS colorTest.parseColor("hsl(30, 12.5%, 12.5%)") is "rgb(35, 31, 27)"
PASS colorTest.parseColor("hsl(60, 12.5%, 12.5%)") is "rgb(35, 35, 27)"
PASS colorTest.parseColor("hsl(90, 12.5%, 12.5%)") is "rgb(31, 35, 27)"
PASS colorTest.parseColor("hsl(120, 12.5%, 12.5%)") is "rgb(27, 35, 27)"
PASS colorTest.parseColor("hsl(150, 12.5%, 12.5%)") is "rgb(27, 35, 31)"
PASS colorTest.parseColor("hsl(180, 12.5%, 12.5%)") is "rgb(27, 35, 35)"
PASS colorTest.parseColor("hsl(210, 12.5%, 12.5%)") is "rgb(27, 31, 35)"
PASS colorTest.parseColor("hsl(240, 12.5%, 12.5%)") is "rgb(27, 27, 35)"
PASS colorTest.parseColor("hsl(270, 12.5%, 12.5%)") is "rgb(31, 27, 35)"
PASS colorTest.parseColor("hsl(300, 12.5%, 12.5%)") is "rgb(35, 27, 35)"
PASS colorTest.parseColor("hsl(330, 12.5%, 12.5%)") is "rgb(35, 27, 31)"
PASS colorTest.parseColor("hsl(0, 25%, 12.5%)") is "rgb(39, 23, 23)"
PASS colorTest.parseColor("hsl(30, 25%, 12.5%)") is "rgb(39, 31, 23)"
PASS colorTest.parseColor("hsl(60, 25%, 12.5%)") is "rgb(39, 39, 23)"
PASS colorTest.parseColor("hsl(90, 25%, 12.5%)") is "rgb(31, 39, 23)"
PASS colorTest.parseColor("hsl(120, 25%, 12.5%)") is "rgb(23, 39, 23)"
PASS colorTest.parseColor("hsl(150, 25%, 12.5%)") is "rgb(23, 39, 31)"
PASS colorTest.parseColor("hsl(180, 25%, 12.5%)") is "rgb(23, 39, 39)"
PASS colorTest.parseColor("hsl(210, 25%, 12.5%)") is "rgb(23, 31, 39)"
PASS colorTest.parseColor("hsl(240, 25%, 12.5%)") is "rgb(23, 23, 39)"
PASS colorTest.parseColor("hsl(270, 25%, 12.5%)") is "rgb(31, 23, 39)"
PASS colorTest.parseColor("hsl(300, 25%, 12.5%)") is "rgb(39, 23, 39)"
PASS colorTest.parseColor("hsl(330, 25%, 12.5%)") is "rgb(39, 23, 31)"
PASS colorTest.parseColor("hsl(0, 37.5%, 12.5%)") is "rgb(43, 19, 19)"
PASS colorTest.parseColor("hsl(30, 37.5%, 12.5%)") is "rgb(43, 31, 19)"
PASS colorTest.parseColor("hsl(60, 37.5%, 12.5%)") is "rgb(43, 43, 19)"
FAIL colorTest.parseColor("hsl(90, 37.5%, 12.5%)") should be rgb(31, 43, 19). Was rgb(32, 43, 19).
PASS colorTest.parseColor("hsl(120, 37.5%, 12.5%)") is "rgb(19, 43, 19)"
FAIL colorTest.parseColor("hsl(150, 37.5%, 12.5%)") should be rgb(19, 43, 31). Was rgb(19, 43, 32).
PASS colorTest.parseColor("hsl(180, 37.5%, 12.5%)") is "rgb(19, 43, 43)"
PASS colorTest.parseColor("hsl(210, 37.5%, 12.5%)") is "rgb(19, 31, 43)"
PASS colorTest.parseColor("hsl(240, 37.5%, 12.5%)") is "rgb(19, 19, 43)"
PASS colorTest.parseColor("hsl(270, 37.5%, 12.5%)") is "rgb(31, 19, 43)"
PASS colorTest.parseColor("hsl(300, 37.5%, 12.5%)") is "rgb(43, 19, 43)"
FAIL colorTest.parseColor("hsl(330, 37.5%, 12.5%)") should be rgb(43, 19, 31). Was rgb(43, 19, 32).
PASS colorTest.parseColor("hsl(0, 50%, 12.5%)") is "rgb(47, 15, 15)"
PASS colorTest.parseColor("hsl(30, 50%, 12.5%)") is "rgb(47, 31, 15)"
PASS colorTest.parseColor("hsl(60, 50%, 12.5%)") is "rgb(47, 47, 15)"
FAIL colorTest.parseColor("hsl(90, 50%, 12.5%)") should be rgb(31, 47, 15). Was rgb(32, 47, 15).
PASS colorTest.parseColor("hsl(120, 50%, 12.5%)") is "rgb(15, 47, 15)"
FAIL colorTest.parseColor("hsl(150, 50%, 12.5%)") should be rgb(15, 47, 31). Was rgb(15, 47, 32).
PASS colorTest.parseColor("hsl(180, 50%, 12.5%)") is "rgb(15, 47, 47)"
PASS colorTest.parseColor("hsl(210, 50%, 12.5%)") is "rgb(15, 31, 47)"
PASS colorTest.parseColor("hsl(240, 50%, 12.5%)") is "rgb(15, 15, 47)"
PASS colorTest.parseColor("hsl(270, 50%, 12.5%)") is "rgb(31, 15, 47)"
PASS colorTest.parseColor("hsl(300, 50%, 12.5%)") is "rgb(47, 15, 47)"
FAIL colorTest.parseColor("hsl(330, 50%, 12.5%)") should be rgb(47, 15, 31). Was rgb(47, 15, 32).
PASS colorTest.parseColor("hsl(0, 62.5%, 12.5%)") is "rgb(51, 11, 11)"
PASS colorTest.parseColor("hsl(30, 62.5%, 12.5%)") is "rgb(51, 31, 11)"
PASS colorTest.parseColor("hsl(60, 62.5%, 12.5%)") is "rgb(51, 51, 11)"
FAIL colorTest.parseColor("hsl(90, 62.5%, 12.5%)") should be rgb(31, 51, 11). Was rgb(32, 51, 11).
PASS colorTest.parseColor("hsl(120, 62.5%, 12.5%)") is "rgb(11, 51, 11)"
FAIL colorTest.parseColor("hsl(150, 62.5%, 12.5%)") should be rgb(11, 51, 31). Was rgb(11, 51, 32).
PASS colorTest.parseColor("hsl(180, 62.5%, 12.5%)") is "rgb(11, 51, 51)"
PASS colorTest.parseColor("hsl(210, 62.5%, 12.5%)") is "rgb(11, 31, 51)"
PASS colorTest.parseColor("hsl(240, 62.5%, 12.5%)") is "rgb(11, 11, 51)"
PASS colorTest.parseColor("hsl(270, 62.5%, 12.5%)") is "rgb(31, 11, 51)"
PASS colorTest.parseColor("hsl(300, 62.5%, 12.5%)") is "rgb(51, 11, 51)"
FAIL colorTest.parseColor("hsl(330, 62.5%, 12.5%)") should be rgb(51, 11, 31). Was rgb(51, 11, 32).
PASS colorTest.parseColor("hsl(0, 75%, 12.5%)") is "rgb(55, 7, 7)"
PASS colorTest.parseColor("hsl(30, 75%, 12.5%)") is "rgb(55, 31, 7)"
PASS colorTest.parseColor("hsl(60, 75%, 12.5%)") is "rgb(55, 55, 7)"
FAIL colorTest.parseColor("hsl(90, 75%, 12.5%)") should be rgb(31, 55, 7). Was rgb(32, 55, 7).
PASS colorTest.parseColor("hsl(120, 75%, 12.5%)") is "rgb(7, 55, 7)"
FAIL colorTest.parseColor("hsl(150, 75%, 12.5%)") should be rgb(7, 55, 31). Was rgb(7, 55, 32).
PASS colorTest.parseColor("hsl(180, 75%, 12.5%)") is "rgb(7, 55, 55)"
PASS colorTest.parseColor("hsl(210, 75%, 12.5%)") is "rgb(7, 31, 55)"
PASS colorTest.parseColor("hsl(240, 75%, 12.5%)") is "rgb(7, 7, 55)"
PASS colorTest.parseColor("hsl(270, 75%, 12.5%)") is "rgb(31, 7, 55)"
PASS colorTest.parseColor("hsl(300, 75%, 12.5%)") is "rgb(55, 7, 55)"
FAIL colorTest.parseColor("hsl(330, 75%, 12.5%)") should be rgb(55, 7, 31). Was rgb(55, 7, 32).
PASS colorTest.parseColor("hsl(0, 87.5%, 12.5%)") is "rgb(59, 3, 3)"
PASS colorTest.parseColor("hsl(30, 87.5%, 12.5%)") is "rgb(59, 31, 3)"
PASS colorTest.parseColor("hsl(60, 87.5%, 12.5%)") is "rgb(59, 59, 3)"
FAIL colorTest.parseColor("hsl(90, 87.5%, 12.5%)") should be rgb(31, 59, 3). Was rgb(32, 59, 3).
PASS colorTest.parseColor("hsl(120, 87.5%, 12.5%)") is "rgb(3, 59, 3)"
FAIL colorTest.parseColor("hsl(150, 87.5%, 12.5%)") should be rgb(3, 59, 31). Was rgb(3, 59, 32).
PASS colorTest.parseColor("hsl(180, 87.5%, 12.5%)") is "rgb(3, 59, 59)"
PASS colorTest.parseColor("hsl(210, 87.5%, 12.5%)") is "rgb(3, 31, 59)"
PASS colorTest.parseColor("hsl(240, 87.5%, 12.5%)") is "rgb(3, 3, 59)"
PASS colorTest.parseColor("hsl(270, 87.5%, 12.5%)") is "rgb(31, 3, 59)"
PASS colorTest.parseColor("hsl(300, 87.5%, 12.5%)") is "rgb(59, 3, 59)"
FAIL colorTest.parseColor("hsl(330, 87.5%, 12.5%)") should be rgb(59, 3, 31). Was rgb(59, 3, 32).
PASS colorTest.parseColor("hsl(0, 100%, 12.5%)") is "rgb(63, 0, 0)"
PASS colorTest.parseColor("hsl(30, 100%, 12.5%)") is "rgb(63, 31, 0)"
PASS colorTest.parseColor("hsl(60, 100%, 12.5%)") is "rgb(63, 63, 0)"
FAIL colorTest.parseColor("hsl(90, 100%, 12.5%)") should be rgb(31, 63, 0). Was rgb(32, 63, 0).
PASS colorTest.parseColor("hsl(120, 100%, 12.5%)") is "rgb(0, 63, 0)"
FAIL colorTest.parseColor("hsl(150, 100%, 12.5%)") should be rgb(0, 63, 31). Was rgb(0, 63, 32).
PASS colorTest.parseColor("hsl(180, 100%, 12.5%)") is "rgb(0, 63, 63)"
PASS colorTest.parseColor("hsl(210, 100%, 12.5%)") is "rgb(0, 31, 63)"
PASS colorTest.parseColor("hsl(240, 100%, 12.5%)") is "rgb(0, 0, 63)"
PASS colorTest.parseColor("hsl(270, 100%, 12.5%)") is "rgb(31, 0, 63)"
PASS colorTest.parseColor("hsl(300, 100%, 12.5%)") is "rgb(63, 0, 63)"
FAIL colorTest.parseColor("hsl(330, 100%, 12.5%)") should be rgb(63, 0, 31). Was rgb(63, 0, 32).
PASS colorTest.parseColor("hsl(0, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(30, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(60, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(90, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(120, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(150, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(180, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(210, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(240, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(270, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(300, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(330, 0%, 25%)") is "rgb(63, 63, 63)"
PASS colorTest.parseColor("hsl(0, 12.5%, 25%)") is "rgb(71, 55, 55)"
PASS colorTest.parseColor("hsl(30, 12.5%, 25%)") is "rgb(71, 63, 55)"
PASS colorTest.parseColor("hsl(60, 12.5%, 25%)") is "rgb(71, 71, 55)"
PASS colorTest.parseColor("hsl(90, 12.5%, 25%)") is "rgb(63, 71, 55)"
PASS colorTest.parseColor("hsl(120, 12.5%, 25%)") is "rgb(55, 71, 55)"
PASS colorTest.parseColor("hsl(150, 12.5%, 25%)") is "rgb(55, 71, 63)"
PASS colorTest.parseColor("hsl(180, 12.5%, 25%)") is "rgb(55, 71, 71)"
PASS colorTest.parseColor("hsl(210, 12.5%, 25%)") is "rgb(55, 63, 71)"
PASS colorTest.parseColor("hsl(240, 12.5%, 25%)") is "rgb(55, 55, 71)"
PASS colorTest.parseColor("hsl(270, 12.5%, 25%)") is "rgb(63, 55, 71)"
PASS colorTest.parseColor("hsl(300, 12.5%, 25%)") is "rgb(71, 55, 71)"
PASS colorTest.parseColor("hsl(330, 12.5%, 25%)") is "rgb(71, 55, 63)"
PASS colorTest.parseColor("hsl(0, 25%, 25%)") is "rgb(79, 47, 47)"
PASS colorTest.parseColor("hsl(30, 25%, 25%)") is "rgb(79, 63, 47)"
PASS colorTest.parseColor("hsl(60, 25%, 25%)") is "rgb(79, 79, 47)"
PASS colorTest.parseColor("hsl(90, 25%, 25%)") is "rgb(63, 79, 47)"
PASS colorTest.parseColor("hsl(120, 25%, 25%)") is "rgb(47, 79, 47)"
PASS colorTest.parseColor("hsl(150, 25%, 25%)") is "rgb(47, 79, 63)"
PASS colorTest.parseColor("hsl(180, 25%, 25%)") is "rgb(47, 79, 79)"
PASS colorTest.parseColor("hsl(210, 25%, 25%)") is "rgb(47, 63, 79)"
PASS colorTest.parseColor("hsl(240, 25%, 25%)") is "rgb(47, 47, 79)"
PASS colorTest.parseColor("hsl(270, 25%, 25%)") is "rgb(63, 47, 79)"
PASS colorTest.parseColor("hsl(300, 25%, 25%)") is "rgb(79, 47, 79)"
PASS colorTest.parseColor("hsl(330, 25%, 25%)") is "rgb(79, 47, 63)"
PASS colorTest.parseColor("hsl(0, 37.5%, 25%)") is "rgb(87, 39, 39)"
PASS colorTest.parseColor("hsl(30, 37.5%, 25%)") is "rgb(87, 63, 39)"
PASS colorTest.parseColor("hsl(60, 37.5%, 25%)") is "rgb(87, 87, 39)"
FAIL colorTest.parseColor("hsl(90, 37.5%, 25%)") should be rgb(63, 87, 39). Was rgb(64, 87, 39).
PASS colorTest.parseColor("hsl(120, 37.5%, 25%)") is "rgb(39, 87, 39)"
FAIL colorTest.parseColor("hsl(150, 37.5%, 25%)") should be rgb(39, 87, 63). Was rgb(39, 87, 64).
PASS colorTest.parseColor("hsl(180, 37.5%, 25%)") is "rgb(39, 87, 87)"
PASS colorTest.parseColor("hsl(210, 37.5%, 25%)") is "rgb(39, 63, 87)"
PASS colorTest.parseColor("hsl(240, 37.5%, 25%)") is "rgb(39, 39, 87)"
PASS colorTest.parseColor("hsl(270, 37.5%, 25%)") is "rgb(63, 39, 87)"
PASS colorTest.parseColor("hsl(300, 37.5%, 25%)") is "rgb(87, 39, 87)"
FAIL colorTest.parseColor("hsl(330, 37.5%, 25%)") should be rgb(87, 39, 63). Was rgb(87, 39, 64).
PASS colorTest.parseColor("hsl(0, 50%, 25%)") is "rgb(95, 31, 31)"
PASS colorTest.parseColor("hsl(30, 50%, 25%)") is "rgb(95, 63, 31)"
PASS colorTest.parseColor("hsl(60, 50%, 25%)") is "rgb(95, 95, 31)"
FAIL colorTest.parseColor("hsl(90, 50%, 25%)") should be rgb(63, 95, 31). Was rgb(64, 95, 31).
PASS colorTest.parseColor("hsl(120, 50%, 25%)") is "rgb(31, 95, 31)"
FAIL colorTest.parseColor("hsl(150, 50%, 25%)") should be rgb(31, 95, 63). Was rgb(31, 95, 64).
PASS colorTest.parseColor("hsl(180, 50%, 25%)") is "rgb(31, 95, 95)"
PASS colorTest.parseColor("hsl(210, 50%, 25%)") is "rgb(31, 63, 95)"
PASS colorTest.parseColor("hsl(240, 50%, 25%)") is "rgb(31, 31, 95)"
PASS colorTest.parseColor("hsl(270, 50%, 25%)") is "rgb(63, 31, 95)"
PASS colorTest.parseColor("hsl(300, 50%, 25%)") is "rgb(95, 31, 95)"
FAIL colorTest.parseColor("hsl(330, 50%, 25%)") should be rgb(95, 31, 63). Was rgb(95, 31, 64).
PASS colorTest.parseColor("hsl(0, 62.5%, 25%)") is "rgb(103, 23, 23)"
PASS colorTest.parseColor("hsl(30, 62.5%, 25%)") is "rgb(103, 63, 23)"
PASS colorTest.parseColor("hsl(60, 62.5%, 25%)") is "rgb(103, 103, 23)"
FAIL colorTest.parseColor("hsl(90, 62.5%, 25%)") should be rgb(63, 103, 23). Was rgb(64, 103, 23).
PASS colorTest.parseColor("hsl(120, 62.5%, 25%)") is "rgb(23, 103, 23)"
FAIL colorTest.parseColor("hsl(150, 62.5%, 25%)") should be rgb(23, 103, 63). Was rgb(23, 103, 64).
PASS colorTest.parseColor("hsl(180, 62.5%, 25%)") is "rgb(23, 103, 103)"
PASS colorTest.parseColor("hsl(210, 62.5%, 25%)") is "rgb(23, 63, 103)"
PASS colorTest.parseColor("hsl(240, 62.5%, 25%)") is "rgb(23, 23, 103)"
PASS colorTest.parseColor("hsl(270, 62.5%, 25%)") is "rgb(63, 23, 103)"
PASS colorTest.parseColor("hsl(300, 62.5%, 25%)") is "rgb(103, 23, 103)"
FAIL colorTest.parseColor("hsl(330, 62.5%, 25%)") should be rgb(103, 23, 63). Was rgb(103, 23, 64).
PASS colorTest.parseColor("hsl(0, 75%, 25%)") is "rgb(111, 15, 15)"
PASS colorTest.parseColor("hsl(30, 75%, 25%)") is "rgb(111, 63, 15)"
PASS colorTest.parseColor("hsl(60, 75%, 25%)") is "rgb(111, 111, 15)"
FAIL colorTest.parseColor("hsl(90, 75%, 25%)") should be rgb(63, 111, 15). Was rgb(64, 111, 15).
PASS colorTest.parseColor("hsl(120, 75%, 25%)") is "rgb(15, 111, 15)"
FAIL colorTest.parseColor("hsl(150, 75%, 25%)") should be rgb(15, 111, 63). Was rgb(15, 111, 64).
PASS colorTest.parseColor("hsl(180, 75%, 25%)") is "rgb(15, 111, 111)"
PASS colorTest.parseColor("hsl(210, 75%, 25%)") is "rgb(15, 63, 111)"
PASS colorTest.parseColor("hsl(240, 75%, 25%)") is "rgb(15, 15, 111)"
PASS colorTest.parseColor("hsl(270, 75%, 25%)") is "rgb(63, 15, 111)"
PASS colorTest.parseColor("hsl(300, 75%, 25%)") is "rgb(111, 15, 111)"
FAIL colorTest.parseColor("hsl(330, 75%, 25%)") should be rgb(111, 15, 63). Was rgb(111, 15, 64).
PASS colorTest.parseColor("hsl(0, 87.5%, 25%)") is "rgb(119, 7, 7)"
PASS colorTest.parseColor("hsl(30, 87.5%, 25%)") is "rgb(119, 63, 7)"
PASS colorTest.parseColor("hsl(60, 87.5%, 25%)") is "rgb(119, 119, 7)"
FAIL colorTest.parseColor("hsl(90, 87.5%, 25%)") should be rgb(63, 119, 7). Was rgb(64, 119, 7).
PASS colorTest.parseColor("hsl(120, 87.5%, 25%)") is "rgb(7, 119, 7)"
FAIL colorTest.parseColor("hsl(150, 87.5%, 25%)") should be rgb(7, 119, 63). Was rgb(7, 119, 64).
PASS colorTest.parseColor("hsl(180, 87.5%, 25%)") is "rgb(7, 119, 119)"
PASS colorTest.parseColor("hsl(210, 87.5%, 25%)") is "rgb(7, 63, 119)"
PASS colorTest.parseColor("hsl(240, 87.5%, 25%)") is "rgb(7, 7, 119)"
PASS colorTest.parseColor("hsl(270, 87.5%, 25%)") is "rgb(63, 7, 119)"
PASS colorTest.parseColor("hsl(300, 87.5%, 25%)") is "rgb(119, 7, 119)"
FAIL colorTest.parseColor("hsl(330, 87.5%, 25%)") should be rgb(119, 7, 63). Was rgb(119, 7, 64).
PASS colorTest.parseColor("hsl(0, 100%, 25%)") is "rgb(127, 0, 0)"
PASS colorTest.parseColor("hsl(30, 100%, 25%)") is "rgb(127, 63, 0)"
PASS colorTest.parseColor("hsl(60, 100%, 25%)") is "rgb(127, 127, 0)"
FAIL colorTest.parseColor("hsl(90, 100%, 25%)") should be rgb(63, 127, 0). Was rgb(64, 127, 0).
PASS colorTest.parseColor("hsl(120, 100%, 25%)") is "rgb(0, 127, 0)"
FAIL colorTest.parseColor("hsl(150, 100%, 25%)") should be rgb(0, 127, 63). Was rgb(0, 127, 64).
PASS colorTest.parseColor("hsl(180, 100%, 25%)") is "rgb(0, 127, 127)"
PASS colorTest.parseColor("hsl(210, 100%, 25%)") is "rgb(0, 63, 127)"
PASS colorTest.parseColor("hsl(240, 100%, 25%)") is "rgb(0, 0, 127)"
PASS colorTest.parseColor("hsl(270, 100%, 25%)") is "rgb(63, 0, 127)"
PASS colorTest.parseColor("hsl(300, 100%, 25%)") is "rgb(127, 0, 127)"
FAIL colorTest.parseColor("hsl(330, 100%, 25%)") should be rgb(127, 0, 63). Was rgb(127, 0, 64).
PASS colorTest.parseColor("hsl(0, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(30, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(60, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(90, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(120, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(150, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(180, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(210, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(240, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(270, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(300, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(330, 0%, 37.5%)") is "rgb(95, 95, 95)"
PASS colorTest.parseColor("hsl(0, 12.5%, 37.5%)") is "rgb(107, 83, 83)"
PASS colorTest.parseColor("hsl(30, 12.5%, 37.5%)") is "rgb(107, 95, 83)"
PASS colorTest.parseColor("hsl(60, 12.5%, 37.5%)") is "rgb(107, 107, 83)"
PASS colorTest.parseColor("hsl(90, 12.5%, 37.5%)") is "rgb(95, 107, 83)"
PASS colorTest.parseColor("hsl(120, 12.5%, 37.5%)") is "rgb(83, 107, 83)"
PASS colorTest.parseColor("hsl(150, 12.5%, 37.5%)") is "rgb(83, 107, 95)"
PASS colorTest.parseColor("hsl(180, 12.5%, 37.5%)") is "rgb(83, 107, 107)"
PASS colorTest.parseColor("hsl(210, 12.5%, 37.5%)") is "rgb(83, 95, 107)"
PASS colorTest.parseColor("hsl(240, 12.5%, 37.5%)") is "rgb(83, 83, 107)"
PASS colorTest.parseColor("hsl(270, 12.5%, 37.5%)") is "rgb(95, 83, 107)"
PASS colorTest.parseColor("hsl(300, 12.5%, 37.5%)") is "rgb(107, 83, 107)"
PASS colorTest.parseColor("hsl(330, 12.5%, 37.5%)") is "rgb(107, 83, 95)"
PASS colorTest.parseColor("hsl(0, 25%, 37.5%)") is "rgb(119, 71, 71)"
PASS colorTest.parseColor("hsl(30, 25%, 37.5%)") is "rgb(119, 95, 71)"
PASS colorTest.parseColor("hsl(60, 25%, 37.5%)") is "rgb(119, 119, 71)"
FAIL colorTest.parseColor("hsl(90, 25%, 37.5%)") should be rgb(95, 119, 71). Was rgb(96, 119, 71).
PASS colorTest.parseColor("hsl(120, 25%, 37.5%)") is "rgb(71, 119, 71)"
FAIL colorTest.parseColor("hsl(150, 25%, 37.5%)") should be rgb(71, 119, 95). Was rgb(71, 119, 96).
PASS colorTest.parseColor("hsl(180, 25%, 37.5%)") is "rgb(71, 119, 119)"
PASS colorTest.parseColor("hsl(210, 25%, 37.5%)") is "rgb(71, 95, 119)"
PASS colorTest.parseColor("hsl(240, 25%, 37.5%)") is "rgb(71, 71, 119)"
PASS colorTest.parseColor("hsl(270, 25%, 37.5%)") is "rgb(95, 71, 119)"
PASS colorTest.parseColor("hsl(300, 25%, 37.5%)") is "rgb(119, 71, 119)"
FAIL colorTest.parseColor("hsl(330, 25%, 37.5%)") should be rgb(119, 71, 95). Was rgb(119, 71, 96).
PASS colorTest.parseColor("hsl(0, 37.5%, 37.5%)") is "rgb(131, 59, 59)"
PASS colorTest.parseColor("hsl(30, 37.5%, 37.5%)") is "rgb(131, 95, 59)"
PASS colorTest.parseColor("hsl(60, 37.5%, 37.5%)") is "rgb(131, 131, 59)"
FAIL colorTest.parseColor("hsl(90, 37.5%, 37.5%)") should be rgb(95, 131, 59). Was rgb(96, 131, 59).
PASS colorTest.parseColor("hsl(120, 37.5%, 37.5%)") is "rgb(59, 131, 59)"
FAIL colorTest.parseColor("hsl(150, 37.5%, 37.5%)") should be rgb(59, 131, 95). Was rgb(59, 131, 96).
PASS colorTest.parseColor("hsl(180, 37.5%, 37.5%)") is "rgb(59, 131, 131)"
PASS colorTest.parseColor("hsl(210, 37.5%, 37.5%)") is "rgb(59, 95, 131)"
PASS colorTest.parseColor("hsl(240, 37.5%, 37.5%)") is "rgb(59, 59, 131)"
PASS colorTest.parseColor("hsl(270, 37.5%, 37.5%)") is "rgb(95, 59, 131)"
PASS colorTest.parseColor("hsl(300, 37.5%, 37.5%)") is "rgb(131, 59, 131)"
FAIL colorTest.parseColor("hsl(330, 37.5%, 37.5%)") should be rgb(131, 59, 95). Was rgb(131, 59, 96).
PASS colorTest.parseColor("hsl(0, 50%, 37.5%)") is "rgb(143, 47, 47)"
PASS colorTest.parseColor("hsl(30, 50%, 37.5%)") is "rgb(143, 95, 47)"
PASS colorTest.parseColor("hsl(60, 50%, 37.5%)") is "rgb(143, 143, 47)"
FAIL colorTest.parseColor("hsl(90, 50%, 37.5%)") should be rgb(95, 143, 47). Was rgb(96, 143, 47).
PASS colorTest.parseColor("hsl(120, 50%, 37.5%)") is "rgb(47, 143, 47)"
FAIL colorTest.parseColor("hsl(150, 50%, 37.5%)") should be rgb(47, 143, 95). Was rgb(47, 143, 96).
PASS colorTest.parseColor("hsl(180, 50%, 37.5%)") is "rgb(47, 143, 143)"
PASS colorTest.parseColor("hsl(210, 50%, 37.5%)") is "rgb(47, 95, 143)"
PASS colorTest.parseColor("hsl(240, 50%, 37.5%)") is "rgb(47, 47, 143)"
PASS colorTest.parseColor("hsl(270, 50%, 37.5%)") is "rgb(95, 47, 143)"
PASS colorTest.parseColor("hsl(300, 50%, 37.5%)") is "rgb(143, 47, 143)"
FAIL colorTest.parseColor("hsl(330, 50%, 37.5%)") should be rgb(143, 47, 95). Was rgb(143, 47, 96).
PASS colorTest.parseColor("hsl(0, 62.5%, 37.5%)") is "rgb(155, 35, 35)"
PASS colorTest.parseColor("hsl(30, 62.5%, 37.5%)") is "rgb(155, 95, 35)"
PASS colorTest.parseColor("hsl(60, 62.5%, 37.5%)") is "rgb(155, 155, 35)"
FAIL colorTest.parseColor("hsl(90, 62.5%, 37.5%)") should be rgb(95, 155, 35). Was rgb(96, 155, 35).
PASS colorTest.parseColor("hsl(120, 62.5%, 37.5%)") is "rgb(35, 155, 35)"
FAIL colorTest.parseColor("hsl(150, 62.5%, 37.5%)") should be rgb(35, 155, 95). Was rgb(35, 155, 96).
PASS colorTest.parseColor("hsl(180, 62.5%, 37.5%)") is "rgb(35, 155, 155)"
PASS colorTest.parseColor("hsl(210, 62.5%, 37.5%)") is "rgb(35, 95, 155)"
PASS colorTest.parseColor("hsl(240, 62.5%, 37.5%)") is "rgb(35, 35, 155)"
PASS colorTest.parseColor("hsl(270, 62.5%, 37.5%)") is "rgb(95, 35, 155)"
PASS colorTest.parseColor("hsl(300, 62.5%, 37.5%)") is "rgb(155, 35, 155)"
FAIL colorTest.parseColor("hsl(330, 62.5%, 37.5%)") should be rgb(155, 35, 95). Was rgb(155, 35, 96).
PASS colorTest.parseColor("hsl(0, 75%, 37.5%)") is "rgb(167, 23, 23)"
PASS colorTest.parseColor("hsl(30, 75%, 37.5%)") is "rgb(167, 95, 23)"
PASS colorTest.parseColor("hsl(60, 75%, 37.5%)") is "rgb(167, 167, 23)"
FAIL colorTest.parseColor("hsl(90, 75%, 37.5%)") should be rgb(95, 167, 23). Was rgb(96, 167, 23).
PASS colorTest.parseColor("hsl(120, 75%, 37.5%)") is "rgb(23, 167, 23)"
FAIL colorTest.parseColor("hsl(150, 75%, 37.5%)") should be rgb(23, 167, 95). Was rgb(23, 167, 96).
PASS colorTest.parseColor("hsl(180, 75%, 37.5%)") is "rgb(23, 167, 167)"
PASS colorTest.parseColor("hsl(210, 75%, 37.5%)") is "rgb(23, 95, 167)"
PASS colorTest.parseColor("hsl(240, 75%, 37.5%)") is "rgb(23, 23, 167)"
PASS colorTest.parseColor("hsl(270, 75%, 37.5%)") is "rgb(95, 23, 167)"
PASS colorTest.parseColor("hsl(300, 75%, 37.5%)") is "rgb(167, 23, 167)"
FAIL colorTest.parseColor("hsl(330, 75%, 37.5%)") should be rgb(167, 23, 95). Was rgb(167, 23, 96).
PASS colorTest.parseColor("hsl(0, 87.5%, 37.5%)") is "rgb(179, 11, 11)"
PASS colorTest.parseColor("hsl(30, 87.5%, 37.5%)") is "rgb(179, 95, 11)"
PASS colorTest.parseColor("hsl(60, 87.5%, 37.5%)") is "rgb(179, 179, 11)"
FAIL colorTest.parseColor("hsl(90, 87.5%, 37.5%)") should be rgb(95, 179, 11). Was rgb(96, 179, 11).
PASS colorTest.parseColor("hsl(120, 87.5%, 37.5%)") is "rgb(11, 179, 11)"
FAIL colorTest.parseColor("hsl(150, 87.5%, 37.5%)") should be rgb(11, 179, 95). Was rgb(11, 179, 96).
PASS colorTest.parseColor("hsl(180, 87.5%, 37.5%)") is "rgb(11, 179, 179)"
PASS colorTest.parseColor("hsl(210, 87.5%, 37.5%)") is "rgb(11, 95, 179)"
PASS colorTest.parseColor("hsl(240, 87.5%, 37.5%)") is "rgb(11, 11, 179)"
PASS colorTest.parseColor("hsl(270, 87.5%, 37.5%)") is "rgb(95, 11, 179)"
PASS colorTest.parseColor("hsl(300, 87.5%, 37.5%)") is "rgb(179, 11, 179)"
FAIL colorTest.parseColor("hsl(330, 87.5%, 37.5%)") should be rgb(179, 11, 95). Was rgb(179, 11, 96).
PASS colorTest.parseColor("hsl(0, 100%, 37.5%)") is "rgb(191, 0, 0)"
PASS colorTest.parseColor("hsl(30, 100%, 37.5%)") is "rgb(191, 95, 0)"
PASS colorTest.parseColor("hsl(60, 100%, 37.5%)") is "rgb(191, 191, 0)"
FAIL colorTest.parseColor("hsl(90, 100%, 37.5%)") should be rgb(95, 191, 0). Was rgb(96, 191, 0).
PASS colorTest.parseColor("hsl(120, 100%, 37.5%)") is "rgb(0, 191, 0)"
FAIL colorTest.parseColor("hsl(150, 100%, 37.5%)") should be rgb(0, 191, 95). Was rgb(0, 191, 96).
PASS colorTest.parseColor("hsl(180, 100%, 37.5%)") is "rgb(0, 191, 191)"
PASS colorTest.parseColor("hsl(210, 100%, 37.5%)") is "rgb(0, 95, 191)"
PASS colorTest.parseColor("hsl(240, 100%, 37.5%)") is "rgb(0, 0, 191)"
PASS colorTest.parseColor("hsl(270, 100%, 37.5%)") is "rgb(95, 0, 191)"
PASS colorTest.parseColor("hsl(300, 100%, 37.5%)") is "rgb(191, 0, 191)"
FAIL colorTest.parseColor("hsl(330, 100%, 37.5%)") should be rgb(191, 0, 95). Was rgb(191, 0, 96).
PASS colorTest.parseColor("hsl(0, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(30, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(60, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(90, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(120, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(150, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(180, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(210, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(240, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(270, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(300, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(330, 0%, 50%)") is "rgb(127, 127, 127)"
PASS colorTest.parseColor("hsl(0, 12.5%, 50%)") is "rgb(143, 111, 111)"
PASS colorTest.parseColor("hsl(30, 12.5%, 50%)") is "rgb(143, 127, 111)"
PASS colorTest.parseColor("hsl(60, 12.5%, 50%)") is "rgb(143, 143, 111)"
PASS colorTest.parseColor("hsl(90, 12.5%, 50%)") is "rgb(127, 143, 111)"
PASS colorTest.parseColor("hsl(120, 12.5%, 50%)") is "rgb(111, 143, 111)"
PASS colorTest.parseColor("hsl(150, 12.5%, 50%)") is "rgb(111, 143, 127)"
PASS colorTest.parseColor("hsl(180, 12.5%, 50%)") is "rgb(111, 143, 143)"
PASS colorTest.parseColor("hsl(210, 12.5%, 50%)") is "rgb(111, 127, 143)"
PASS colorTest.parseColor("hsl(240, 12.5%, 50%)") is "rgb(111, 111, 143)"
PASS colorTest.parseColor("hsl(270, 12.5%, 50%)") is "rgb(127, 111, 143)"
PASS colorTest.parseColor("hsl(300, 12.5%, 50%)") is "rgb(143, 111, 143)"
PASS colorTest.parseColor("hsl(330, 12.5%, 50%)") is "rgb(143, 111, 127)"
PASS colorTest.parseColor("hsl(0, 25%, 50%)") is "rgb(159, 95, 95)"
PASS colorTest.parseColor("hsl(30, 25%, 50%)") is "rgb(159, 127, 95)"
PASS colorTest.parseColor("hsl(60, 25%, 50%)") is "rgb(159, 159, 95)"
PASS colorTest.parseColor("hsl(90, 25%, 50%)") is "rgb(127, 159, 95)"
PASS colorTest.parseColor("hsl(120, 25%, 50%)") is "rgb(95, 159, 95)"
PASS colorTest.parseColor("hsl(150, 25%, 50%)") is "rgb(95, 159, 127)"
PASS colorTest.parseColor("hsl(180, 25%, 50%)") is "rgb(95, 159, 159)"
PASS colorTest.parseColor("hsl(210, 25%, 50%)") is "rgb(95, 127, 159)"
PASS colorTest.parseColor("hsl(240, 25%, 50%)") is "rgb(95, 95, 159)"
PASS colorTest.parseColor("hsl(270, 25%, 50%)") is "rgb(127, 95, 159)"
PASS colorTest.parseColor("hsl(300, 25%, 50%)") is "rgb(159, 95, 159)"
PASS colorTest.parseColor("hsl(330, 25%, 50%)") is "rgb(159, 95, 127)"
PASS colorTest.parseColor("hsl(0, 37.5%, 50%)") is "rgb(175, 79, 79)"
PASS colorTest.parseColor("hsl(30, 37.5%, 50%)") is "rgb(175, 127, 79)"
PASS colorTest.parseColor("hsl(60, 37.5%, 50%)") is "rgb(175, 175, 79)"
FAIL colorTest.parseColor("hsl(90, 37.5%, 50%)") should be rgb(127, 175, 79). Was rgb(128, 175, 79).
PASS colorTest.parseColor("hsl(120, 37.5%, 50%)") is "rgb(79, 175, 79)"
FAIL colorTest.parseColor("hsl(150, 37.5%, 50%)") should be rgb(79, 175, 127). Was rgb(79, 175, 128).
PASS colorTest.parseColor("hsl(180, 37.5%, 50%)") is "rgb(79, 175, 175)"
PASS colorTest.parseColor("hsl(210, 37.5%, 50%)") is "rgb(79, 127, 175)"
PASS colorTest.parseColor("hsl(240, 37.5%, 50%)") is "rgb(79, 79, 175)"
PASS colorTest.parseColor("hsl(270, 37.5%, 50%)") is "rgb(127, 79, 175)"
PASS colorTest.parseColor("hsl(300, 37.5%, 50%)") is "rgb(175, 79, 175)"
FAIL colorTest.parseColor("hsl(330, 37.5%, 50%)") should be rgb(175, 79, 127). Was rgb(175, 79, 128).
PASS colorTest.parseColor("hsl(0, 50%, 50%)") is "rgb(191, 63, 63)"
PASS colorTest.parseColor("hsl(30, 50%, 50%)") is "rgb(191, 127, 63)"
PASS colorTest.parseColor("hsl(60, 50%, 50%)") is "rgb(191, 191, 63)"
FAIL colorTest.parseColor("hsl(90, 50%, 50%)") should be rgb(127, 191, 63). Was rgb(128, 191, 63).
PASS colorTest.parseColor("hsl(120, 50%, 50%)") is "rgb(63, 191, 63)"
FAIL colorTest.parseColor("hsl(150, 50%, 50%)") should be rgb(63, 191, 127). Was rgb(63, 191, 128).
PASS colorTest.parseColor("hsl(180, 50%, 50%)") is "rgb(63, 191, 191)"
PASS colorTest.parseColor("hsl(210, 50%, 50%)") is "rgb(63, 127, 191)"
PASS colorTest.parseColor("hsl(240, 50%, 50%)") is "rgb(63, 63, 191)"
PASS colorTest.parseColor("hsl(270, 50%, 50%)") is "rgb(127, 63, 191)"
PASS colorTest.parseColor("hsl(300, 50%, 50%)") is "rgb(191, 63, 191)"
FAIL colorTest.parseColor("hsl(330, 50%, 50%)") should be rgb(191, 63, 127). Was rgb(191, 63, 128).
PASS colorTest.parseColor("hsl(0, 62.5%, 50%)") is "rgb(207, 47, 47)"
PASS colorTest.parseColor("hsl(30, 62.5%, 50%)") is "rgb(207, 127, 47)"
PASS colorTest.parseColor("hsl(60, 62.5%, 50%)") is "rgb(207, 207, 47)"
FAIL colorTest.parseColor("hsl(90, 62.5%, 50%)") should be rgb(127, 207, 47). Was rgb(128, 207, 47).
PASS colorTest.parseColor("hsl(120, 62.5%, 50%)") is "rgb(47, 207, 47)"
FAIL colorTest.parseColor("hsl(150, 62.5%, 50%)") should be rgb(47, 207, 127). Was rgb(47, 207, 128).
PASS colorTest.parseColor("hsl(180, 62.5%, 50%)") is "rgb(47, 207, 207)"
PASS colorTest.parseColor("hsl(210, 62.5%, 50%)") is "rgb(47, 127, 207)"
PASS colorTest.parseColor("hsl(240, 62.5%, 50%)") is "rgb(47, 47, 207)"
PASS colorTest.parseColor("hsl(270, 62.5%, 50%)") is "rgb(127, 47, 207)"
PASS colorTest.parseColor("hsl(300, 62.5%, 50%)") is "rgb(207, 47, 207)"
FAIL colorTest.parseColor("hsl(330, 62.5%, 50%)") should be rgb(207, 47, 127). Was rgb(207, 47, 128).
PASS colorTest.parseColor("hsl(0, 75%, 50%)") is "rgb(223, 31, 31)"
PASS colorTest.parseColor("hsl(30, 75%, 50%)") is "rgb(223, 127, 31)"
PASS colorTest.parseColor("hsl(60, 75%, 50%)") is "rgb(223, 223, 31)"
FAIL colorTest.parseColor("hsl(90, 75%, 50%)") should be rgb(127, 223, 31). Was rgb(128, 223, 31).
PASS colorTest.parseColor("hsl(120, 75%, 50%)") is "rgb(31, 223, 31)"
FAIL colorTest.parseColor("hsl(150, 75%, 50%)") should be rgb(31, 223, 127). Was rgb(31, 223, 128).
PASS colorTest.parseColor("hsl(180, 75%, 50%)") is "rgb(31, 223, 223)"
PASS colorTest.parseColor("hsl(210, 75%, 50%)") is "rgb(31, 127, 223)"
PASS colorTest.parseColor("hsl(240, 75%, 50%)") is "rgb(31, 31, 223)"
PASS colorTest.parseColor("hsl(270, 75%, 50%)") is "rgb(127, 31, 223)"
PASS colorTest.parseColor("hsl(300, 75%, 50%)") is "rgb(223, 31, 223)"
FAIL colorTest.parseColor("hsl(330, 75%, 50%)") should be rgb(223, 31, 127). Was rgb(223, 31, 128).
PASS colorTest.parseColor("hsl(0, 87.5%, 50%)") is "rgb(239, 15, 15)"
PASS colorTest.parseColor("hsl(30, 87.5%, 50%)") is "rgb(239, 127, 15)"
PASS colorTest.parseColor("hsl(60, 87.5%, 50%)") is "rgb(239, 239, 15)"
FAIL colorTest.parseColor("hsl(90, 87.5%, 50%)") should be rgb(127, 239, 15). Was rgb(128, 239, 15).
PASS colorTest.parseColor("hsl(120, 87.5%, 50%)") is "rgb(15, 239, 15)"
FAIL colorTest.parseColor("hsl(150, 87.5%, 50%)") should be rgb(15, 239, 127). Was rgb(15, 239, 128).
PASS colorTest.parseColor("hsl(180, 87.5%, 50%)") is "rgb(15, 239, 239)"
PASS colorTest.parseColor("hsl(210, 87.5%, 50%)") is "rgb(15, 127, 239)"
PASS colorTest.parseColor("hsl(240, 87.5%, 50%)") is "rgb(15, 15, 239)"
PASS colorTest.parseColor("hsl(270, 87.5%, 50%)") is "rgb(127, 15, 239)"
PASS colorTest.parseColor("hsl(300, 87.5%, 50%)") is "rgb(239, 15, 239)"
FAIL colorTest.parseColor("hsl(330, 87.5%, 50%)") should be rgb(239, 15, 127). Was rgb(239, 15, 128).
PASS colorTest.parseColor("hsl(0, 100%, 50%)") is "rgb(255, 0, 0)"
PASS colorTest.parseColor("hsl(30, 100%, 50%)") is "rgb(255, 127, 0)"
PASS colorTest.parseColor("hsl(60, 100%, 50%)") is "rgb(255, 255, 0)"
FAIL colorTest.parseColor("hsl(90, 100%, 50%)") should be rgb(127, 255, 0). Was rgb(128, 255, 0).
PASS colorTest.parseColor("hsl(120, 100%, 50%)") is "rgb(0, 255, 0)"
FAIL colorTest.parseColor("hsl(150, 100%, 50%)") should be rgb(0, 255, 127). Was rgb(0, 255, 128).
PASS colorTest.parseColor("hsl(180, 100%, 50%)") is "rgb(0, 255, 255)"
PASS colorTest.parseColor("hsl(210, 100%, 50%)") is "rgb(0, 127, 255)"
PASS colorTest.parseColor("hsl(240, 100%, 50%)") is "rgb(0, 0, 255)"
PASS colorTest.parseColor("hsl(270, 100%, 50%)") is "rgb(127, 0, 255)"
PASS colorTest.parseColor("hsl(300, 100%, 50%)") is "rgb(255, 0, 255)"
FAIL colorTest.parseColor("hsl(330, 100%, 50%)") should be rgb(255, 0, 127). Was rgb(255, 0, 128).
PASS colorTest.parseColor("hsl(0, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(30, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(60, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(90, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(120, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(150, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(180, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(210, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(240, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(270, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(300, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(330, 0%, 62.5%)") is "rgb(159, 159, 159)"
PASS colorTest.parseColor("hsl(0, 12.5%, 62.5%)") is "rgb(171, 147, 147)"
PASS colorTest.parseColor("hsl(30, 12.5%, 62.5%)") is "rgb(171, 159, 147)"
PASS colorTest.parseColor("hsl(60, 12.5%, 62.5%)") is "rgb(171, 171, 147)"
PASS colorTest.parseColor("hsl(90, 12.5%, 62.5%)") is "rgb(159, 171, 147)"
PASS colorTest.parseColor("hsl(120, 12.5%, 62.5%)") is "rgb(147, 171, 147)"
PASS colorTest.parseColor("hsl(150, 12.5%, 62.5%)") is "rgb(147, 171, 159)"
PASS colorTest.parseColor("hsl(180, 12.5%, 62.5%)") is "rgb(147, 171, 171)"
PASS colorTest.parseColor("hsl(210, 12.5%, 62.5%)") is "rgb(147, 159, 171)"
PASS colorTest.parseColor("hsl(240, 12.5%, 62.5%)") is "rgb(147, 147, 171)"
PASS colorTest.parseColor("hsl(270, 12.5%, 62.5%)") is "rgb(159, 147, 171)"
PASS colorTest.parseColor("hsl(300, 12.5%, 62.5%)") is "rgb(171, 147, 171)"
PASS colorTest.parseColor("hsl(330, 12.5%, 62.5%)") is "rgb(171, 147, 159)"
PASS colorTest.parseColor("hsl(0, 25%, 62.5%)") is "rgb(183, 135, 135)"
PASS colorTest.parseColor("hsl(30, 25%, 62.5%)") is "rgb(183, 159, 135)"
PASS colorTest.parseColor("hsl(60, 25%, 62.5%)") is "rgb(183, 183, 135)"
PASS colorTest.parseColor("hsl(90, 25%, 62.5%)") is "rgb(159, 183, 135)"
PASS colorTest.parseColor("hsl(120, 25%, 62.5%)") is "rgb(135, 183, 135)"
PASS colorTest.parseColor("hsl(150, 25%, 62.5%)") is "rgb(135, 183, 159)"
PASS colorTest.parseColor("hsl(180, 25%, 62.5%)") is "rgb(135, 183, 183)"
PASS colorTest.parseColor("hsl(210, 25%, 62.5%)") is "rgb(135, 159, 183)"
PASS colorTest.parseColor("hsl(240, 25%, 62.5%)") is "rgb(135, 135, 183)"
PASS colorTest.parseColor("hsl(270, 25%, 62.5%)") is "rgb(159, 135, 183)"
PASS colorTest.parseColor("hsl(300, 25%, 62.5%)") is "rgb(183, 135, 183)"
PASS colorTest.parseColor("hsl(330, 25%, 62.5%)") is "rgb(183, 135, 159)"
PASS colorTest.parseColor("hsl(0, 37.5%, 62.5%)") is "rgb(195, 123, 123)"
PASS colorTest.parseColor("hsl(30, 37.5%, 62.5%)") is "rgb(195, 159, 123)"
PASS colorTest.parseColor("hsl(60, 37.5%, 62.5%)") is "rgb(195, 195, 123)"
PASS colorTest.parseColor("hsl(90, 37.5%, 62.5%)") is "rgb(159, 195, 123)"
PASS colorTest.parseColor("hsl(120, 37.5%, 62.5%)") is "rgb(123, 195, 123)"
PASS colorTest.parseColor("hsl(150, 37.5%, 62.5%)") is "rgb(123, 195, 159)"
PASS colorTest.parseColor("hsl(180, 37.5%, 62.5%)") is "rgb(123, 195, 195)"
PASS colorTest.parseColor("hsl(210, 37.5%, 62.5%)") is "rgb(123, 159, 195)"
PASS colorTest.parseColor("hsl(240, 37.5%, 62.5%)") is "rgb(123, 123, 195)"
PASS colorTest.parseColor("hsl(270, 37.5%, 62.5%)") is "rgb(159, 123, 195)"
PASS colorTest.parseColor("hsl(300, 37.5%, 62.5%)") is "rgb(195, 123, 195)"
PASS colorTest.parseColor("hsl(330, 37.5%, 62.5%)") is "rgb(195, 123, 159)"
PASS colorTest.parseColor("hsl(0, 50%, 62.5%)") is "rgb(207, 111, 111)"
PASS colorTest.parseColor("hsl(30, 50%, 62.5%)") is "rgb(207, 159, 111)"
PASS colorTest.parseColor("hsl(60, 50%, 62.5%)") is "rgb(207, 207, 111)"
FAIL colorTest.parseColor("hsl(90, 50%, 62.5%)") should be rgb(159, 207, 111). Was rgb(160, 207, 111).
PASS colorTest.parseColor("hsl(120, 50%, 62.5%)") is "rgb(111, 207, 111)"
FAIL colorTest.parseColor("hsl(150, 50%, 62.5%)") should be rgb(111, 207, 159). Was rgb(111, 207, 160).
PASS colorTest.parseColor("hsl(180, 50%, 62.5%)") is "rgb(111, 207, 207)"
PASS colorTest.parseColor("hsl(210, 50%, 62.5%)") is "rgb(111, 159, 207)"
PASS colorTest.parseColor("hsl(240, 50%, 62.5%)") is "rgb(111, 111, 207)"
PASS colorTest.parseColor("hsl(270, 50%, 62.5%)") is "rgb(159, 111, 207)"
PASS colorTest.parseColor("hsl(300, 50%, 62.5%)") is "rgb(207, 111, 207)"
FAIL colorTest.parseColor("hsl(330, 50%, 62.5%)") should be rgb(207, 111, 159). Was rgb(207, 111, 160).
PASS colorTest.parseColor("hsl(0, 62.5%, 62.5%)") is "rgb(219, 99, 99)"
PASS colorTest.parseColor("hsl(30, 62.5%, 62.5%)") is "rgb(219, 159, 99)"
PASS colorTest.parseColor("hsl(60, 62.5%, 62.5%)") is "rgb(219, 219, 99)"
FAIL colorTest.parseColor("hsl(90, 62.5%, 62.5%)") should be rgb(159, 219, 99). Was rgb(160, 219, 99).
PASS colorTest.parseColor("hsl(120, 62.5%, 62.5%)") is "rgb(99, 219, 99)"
FAIL colorTest.parseColor("hsl(150, 62.5%, 62.5%)") should be rgb(99, 219, 159). Was rgb(99, 219, 160).
PASS colorTest.parseColor("hsl(180, 62.5%, 62.5%)") is "rgb(99, 219, 219)"
PASS colorTest.parseColor("hsl(210, 62.5%, 62.5%)") is "rgb(99, 159, 219)"
PASS colorTest.parseColor("hsl(240, 62.5%, 62.5%)") is "rgb(99, 99, 219)"
PASS colorTest.parseColor("hsl(270, 62.5%, 62.5%)") is "rgb(159, 99, 219)"
PASS colorTest.parseColor("hsl(300, 62.5%, 62.5%)") is "rgb(219, 99, 219)"
FAIL colorTest.parseColor("hsl(330, 62.5%, 62.5%)") should be rgb(219, 99, 159). Was rgb(219, 99, 160).
PASS colorTest.parseColor("hsl(0, 75%, 62.5%)") is "rgb(231, 87, 87)"
PASS colorTest.parseColor("hsl(30, 75%, 62.5%)") is "rgb(231, 159, 87)"
PASS colorTest.parseColor("hsl(60, 75%, 62.5%)") is "rgb(231, 231, 87)"
FAIL colorTest.parseColor("hsl(90, 75%, 62.5%)") should be rgb(159, 231, 87). Was rgb(160, 231, 87).
PASS colorTest.parseColor("hsl(120, 75%, 62.5%)") is "rgb(87, 231, 87)"
FAIL colorTest.parseColor("hsl(150, 75%, 62.5%)") should be rgb(87, 231, 159). Was rgb(87, 231, 160).
PASS colorTest.parseColor("hsl(180, 75%, 62.5%)") is "rgb(87, 231, 231)"
PASS colorTest.parseColor("hsl(210, 75%, 62.5%)") is "rgb(87, 159, 231)"
PASS colorTest.parseColor("hsl(240, 75%, 62.5%)") is "rgb(87, 87, 231)"
PASS colorTest.parseColor("hsl(270, 75%, 62.5%)") is "rgb(159, 87, 231)"
PASS colorTest.parseColor("hsl(300, 75%, 62.5%)") is "rgb(231, 87, 231)"
FAIL colorTest.parseColor("hsl(330, 75%, 62.5%)") should be rgb(231, 87, 159). Was rgb(231, 87, 160).
PASS colorTest.parseColor("hsl(0, 87.5%, 62.5%)") is "rgb(243, 75, 75)"
PASS colorTest.parseColor("hsl(30, 87.5%, 62.5%)") is "rgb(243, 159, 75)"
PASS colorTest.parseColor("hsl(60, 87.5%, 62.5%)") is "rgb(243, 243, 75)"
FAIL colorTest.parseColor("hsl(90, 87.5%, 62.5%)") should be rgb(159, 243, 75). Was rgb(160, 243, 75).
PASS colorTest.parseColor("hsl(120, 87.5%, 62.5%)") is "rgb(75, 243, 75)"
FAIL colorTest.parseColor("hsl(150, 87.5%, 62.5%)") should be rgb(75, 243, 159). Was rgb(75, 243, 160).
PASS colorTest.parseColor("hsl(180, 87.5%, 62.5%)") is "rgb(75, 243, 243)"
PASS colorTest.parseColor("hsl(210, 87.5%, 62.5%)") is "rgb(75, 159, 243)"
PASS colorTest.parseColor("hsl(240, 87.5%, 62.5%)") is "rgb(75, 75, 243)"
PASS colorTest.parseColor("hsl(270, 87.5%, 62.5%)") is "rgb(159, 75, 243)"
PASS colorTest.parseColor("hsl(300, 87.5%, 62.5%)") is "rgb(243, 75, 243)"
FAIL colorTest.parseColor("hsl(330, 87.5%, 62.5%)") should be rgb(243, 75, 159). Was rgb(243, 75, 160).
PASS colorTest.parseColor("hsl(0, 100%, 62.5%)") is "rgb(255, 63, 63)"
PASS colorTest.parseColor("hsl(30, 100%, 62.5%)") is "rgb(255, 159, 63)"
PASS colorTest.parseColor("hsl(60, 100%, 62.5%)") is "rgb(255, 255, 63)"
FAIL colorTest.parseColor("hsl(90, 100%, 62.5%)") should be rgb(159, 255, 63). Was rgb(160, 255, 63).
PASS colorTest.parseColor("hsl(120, 100%, 62.5%)") is "rgb(63, 255, 63)"
FAIL colorTest.parseColor("hsl(150, 100%, 62.5%)") should be rgb(63, 255, 159). Was rgb(63, 255, 160).
PASS colorTest.parseColor("hsl(180, 100%, 62.5%)") is "rgb(63, 255, 255)"
PASS colorTest.parseColor("hsl(210, 100%, 62.5%)") is "rgb(63, 159, 255)"
PASS colorTest.parseColor("hsl(240, 100%, 62.5%)") is "rgb(63, 63, 255)"
PASS colorTest.parseColor("hsl(270, 100%, 62.5%)") is "rgb(159, 63, 255)"
PASS colorTest.parseColor("hsl(300, 100%, 62.5%)") is "rgb(255, 63, 255)"
FAIL colorTest.parseColor("hsl(330, 100%, 62.5%)") should be rgb(255, 63, 159). Was rgb(255, 63, 160).
PASS colorTest.parseColor("hsl(0, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(30, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(60, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(90, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(120, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(150, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(180, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(210, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(240, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(270, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(300, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(330, 0%, 75%)") is "rgb(191, 191, 191)"
PASS colorTest.parseColor("hsl(0, 12.5%, 75%)") is "rgb(199, 183, 183)"
PASS colorTest.parseColor("hsl(30, 12.5%, 75%)") is "rgb(199, 191, 183)"
PASS colorTest.parseColor("hsl(60, 12.5%, 75%)") is "rgb(199, 199, 183)"
PASS colorTest.parseColor("hsl(90, 12.5%, 75%)") is "rgb(191, 199, 183)"
PASS colorTest.parseColor("hsl(120, 12.5%, 75%)") is "rgb(183, 199, 183)"
PASS colorTest.parseColor("hsl(150, 12.5%, 75%)") is "rgb(183, 199, 191)"
PASS colorTest.parseColor("hsl(180, 12.5%, 75%)") is "rgb(183, 199, 199)"
PASS colorTest.parseColor("hsl(210, 12.5%, 75%)") is "rgb(183, 191, 199)"
PASS colorTest.parseColor("hsl(240, 12.5%, 75%)") is "rgb(183, 183, 199)"
PASS colorTest.parseColor("hsl(270, 12.5%, 75%)") is "rgb(191, 183, 199)"
PASS colorTest.parseColor("hsl(300, 12.5%, 75%)") is "rgb(199, 183, 199)"
PASS colorTest.parseColor("hsl(330, 12.5%, 75%)") is "rgb(199, 183, 191)"
PASS colorTest.parseColor("hsl(0, 25%, 75%)") is "rgb(207, 175, 175)"
PASS colorTest.parseColor("hsl(30, 25%, 75%)") is "rgb(207, 191, 175)"
PASS colorTest.parseColor("hsl(60, 25%, 75%)") is "rgb(207, 207, 175)"
PASS colorTest.parseColor("hsl(90, 25%, 75%)") is "rgb(191, 207, 175)"
PASS colorTest.parseColor("hsl(120, 25%, 75%)") is "rgb(175, 207, 175)"
PASS colorTest.parseColor("hsl(150, 25%, 75%)") is "rgb(175, 207, 191)"
PASS colorTest.parseColor("hsl(180, 25%, 75%)") is "rgb(175, 207, 207)"
PASS colorTest.parseColor("hsl(210, 25%, 75%)") is "rgb(175, 191, 207)"
PASS colorTest.parseColor("hsl(240, 25%, 75%)") is "rgb(175, 175, 207)"
PASS colorTest.parseColor("hsl(270, 25%, 75%)") is "rgb(191, 175, 207)"
PASS colorTest.parseColor("hsl(300, 25%, 75%)") is "rgb(207, 175, 207)"
PASS colorTest.parseColor("hsl(330, 25%, 75%)") is "rgb(207, 175, 191)"
PASS colorTest.parseColor("hsl(0, 37.5%, 75%)") is "rgb(215, 167, 167)"
PASS colorTest.parseColor("hsl(30, 37.5%, 75%)") is "rgb(215, 191, 167)"
PASS colorTest.parseColor("hsl(60, 37.5%, 75%)") is "rgb(215, 215, 167)"
PASS colorTest.parseColor("hsl(90, 37.5%, 75%)") is "rgb(191, 215, 167)"
PASS colorTest.parseColor("hsl(120, 37.5%, 75%)") is "rgb(167, 215, 167)"
PASS colorTest.parseColor("hsl(150, 37.5%, 75%)") is "rgb(167, 215, 191)"
PASS colorTest.parseColor("hsl(180, 37.5%, 75%)") is "rgb(167, 215, 215)"
PASS colorTest.parseColor("hsl(210, 37.5%, 75%)") is "rgb(167, 191, 215)"
PASS colorTest.parseColor("hsl(240, 37.5%, 75%)") is "rgb(167, 167, 215)"
PASS colorTest.parseColor("hsl(270, 37.5%, 75%)") is "rgb(191, 167, 215)"
PASS colorTest.parseColor("hsl(300, 37.5%, 75%)") is "rgb(215, 167, 215)"
PASS colorTest.parseColor("hsl(330, 37.5%, 75%)") is "rgb(215, 167, 191)"
PASS colorTest.parseColor("hsl(0, 50%, 75%)") is "rgb(223, 159, 159)"
PASS colorTest.parseColor("hsl(30, 50%, 75%)") is "rgb(223, 191, 159)"
PASS colorTest.parseColor("hsl(60, 50%, 75%)") is "rgb(223, 223, 159)"
PASS colorTest.parseColor("hsl(90, 50%, 75%)") is "rgb(191, 223, 159)"
PASS colorTest.parseColor("hsl(120, 50%, 75%)") is "rgb(159, 223, 159)"
PASS colorTest.parseColor("hsl(150, 50%, 75%)") is "rgb(159, 223, 191)"
PASS colorTest.parseColor("hsl(180, 50%, 75%)") is "rgb(159, 223, 223)"
PASS colorTest.parseColor("hsl(210, 50%, 75%)") is "rgb(159, 191, 223)"
PASS colorTest.parseColor("hsl(240, 50%, 75%)") is "rgb(159, 159, 223)"
PASS colorTest.parseColor("hsl(270, 50%, 75%)") is "rgb(191, 159, 223)"
PASS colorTest.parseColor("hsl(300, 50%, 75%)") is "rgb(223, 159, 223)"
PASS colorTest.parseColor("hsl(330, 50%, 75%)") is "rgb(223, 159, 191)"
PASS colorTest.parseColor("hsl(0, 62.5%, 75%)") is "rgb(231, 151, 151)"
PASS colorTest.parseColor("hsl(30, 62.5%, 75%)") is "rgb(231, 191, 151)"
PASS colorTest.parseColor("hsl(60, 62.5%, 75%)") is "rgb(231, 231, 151)"
PASS colorTest.parseColor("hsl(90, 62.5%, 75%)") is "rgb(191, 231, 151)"
PASS colorTest.parseColor("hsl(120, 62.5%, 75%)") is "rgb(151, 231, 151)"
PASS colorTest.parseColor("hsl(150, 62.5%, 75%)") is "rgb(151, 231, 191)"
PASS colorTest.parseColor("hsl(180, 62.5%, 75%)") is "rgb(151, 231, 231)"
PASS colorTest.parseColor("hsl(210, 62.5%, 75%)") is "rgb(151, 191, 231)"
PASS colorTest.parseColor("hsl(240, 62.5%, 75%)") is "rgb(151, 151, 231)"
PASS colorTest.parseColor("hsl(270, 62.5%, 75%)") is "rgb(191, 151, 231)"
PASS colorTest.parseColor("hsl(300, 62.5%, 75%)") is "rgb(231, 151, 231)"
PASS colorTest.parseColor("hsl(330, 62.5%, 75%)") is "rgb(231, 151, 191)"
PASS colorTest.parseColor("hsl(0, 75%, 75%)") is "rgb(239, 143, 143)"
PASS colorTest.parseColor("hsl(30, 75%, 75%)") is "rgb(239, 191, 143)"
PASS colorTest.parseColor("hsl(60, 75%, 75%)") is "rgb(239, 239, 143)"
FAIL colorTest.parseColor("hsl(90, 75%, 75%)") should be rgb(191, 239, 143). Was rgb(192, 239, 143).
PASS colorTest.parseColor("hsl(120, 75%, 75%)") is "rgb(143, 239, 143)"
FAIL colorTest.parseColor("hsl(150, 75%, 75%)") should be rgb(143, 239, 191). Was rgb(143, 239, 192).
PASS colorTest.parseColor("hsl(180, 75%, 75%)") is "rgb(143, 239, 239)"
PASS colorTest.parseColor("hsl(210, 75%, 75%)") is "rgb(143, 191, 239)"
PASS colorTest.parseColor("hsl(240, 75%, 75%)") is "rgb(143, 143, 239)"
PASS colorTest.parseColor("hsl(270, 75%, 75%)") is "rgb(191, 143, 239)"
PASS colorTest.parseColor("hsl(300, 75%, 75%)") is "rgb(239, 143, 239)"
FAIL colorTest.parseColor("hsl(330, 75%, 75%)") should be rgb(239, 143, 191). Was rgb(239, 143, 192).
PASS colorTest.parseColor("hsl(0, 87.5%, 75%)") is "rgb(247, 135, 135)"
PASS colorTest.parseColor("hsl(30, 87.5%, 75%)") is "rgb(247, 191, 135)"
PASS colorTest.parseColor("hsl(60, 87.5%, 75%)") is "rgb(247, 247, 135)"
FAIL colorTest.parseColor("hsl(90, 87.5%, 75%)") should be rgb(191, 247, 135). Was rgb(192, 247, 135).
PASS colorTest.parseColor("hsl(120, 87.5%, 75%)") is "rgb(135, 247, 135)"
FAIL colorTest.parseColor("hsl(150, 87.5%, 75%)") should be rgb(135, 247, 191). Was rgb(135, 247, 192).
PASS colorTest.parseColor("hsl(180, 87.5%, 75%)") is "rgb(135, 247, 247)"
PASS colorTest.parseColor("hsl(210, 87.5%, 75%)") is "rgb(135, 191, 247)"
PASS colorTest.parseColor("hsl(240, 87.5%, 75%)") is "rgb(135, 135, 247)"
PASS colorTest.parseColor("hsl(270, 87.5%, 75%)") is "rgb(191, 135, 247)"
PASS colorTest.parseColor("hsl(300, 87.5%, 75%)") is "rgb(247, 135, 247)"
FAIL colorTest.parseColor("hsl(330, 87.5%, 75%)") should be rgb(247, 135, 191). Was rgb(247, 135, 192).
PASS colorTest.parseColor("hsl(0, 100%, 75%)") is "rgb(255, 127, 127)"
PASS colorTest.parseColor("hsl(30, 100%, 75%)") is "rgb(255, 191, 127)"
PASS colorTest.parseColor("hsl(60, 100%, 75%)") is "rgb(255, 255, 127)"
FAIL colorTest.parseColor("hsl(90, 100%, 75%)") should be rgb(191, 255, 127). Was rgb(192, 255, 127).
PASS colorTest.parseColor("hsl(120, 100%, 75%)") is "rgb(127, 255, 127)"
FAIL colorTest.parseColor("hsl(150, 100%, 75%)") should be rgb(127, 255, 191). Was rgb(127, 255, 192).
PASS colorTest.parseColor("hsl(180, 100%, 75%)") is "rgb(127, 255, 255)"
PASS colorTest.parseColor("hsl(210, 100%, 75%)") is "rgb(127, 191, 255)"
PASS colorTest.parseColor("hsl(240, 100%, 75%)") is "rgb(127, 127, 255)"
PASS colorTest.parseColor("hsl(270, 100%, 75%)") is "rgb(191, 127, 255)"
PASS colorTest.parseColor("hsl(300, 100%, 75%)") is "rgb(255, 127, 255)"
FAIL colorTest.parseColor("hsl(330, 100%, 75%)") should be rgb(255, 127, 191). Was rgb(255, 127, 192).
PASS colorTest.parseColor("hsl(0, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(30, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(60, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(90, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(120, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(150, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(180, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(210, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(240, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(270, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(300, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(330, 0%, 87.5%)") is "rgb(223, 223, 223)"
PASS colorTest.parseColor("hsl(0, 12.5%, 87.5%)") is "rgb(227, 219, 219)"
PASS colorTest.parseColor("hsl(30, 12.5%, 87.5%)") is "rgb(227, 223, 219)"
PASS colorTest.parseColor("hsl(60, 12.5%, 87.5%)") is "rgb(227, 227, 219)"
PASS colorTest.parseColor("hsl(90, 12.5%, 87.5%)") is "rgb(223, 227, 219)"
PASS colorTest.parseColor("hsl(120, 12.5%, 87.5%)") is "rgb(219, 227, 219)"
PASS colorTest.parseColor("hsl(150, 12.5%, 87.5%)") is "rgb(219, 227, 223)"
PASS colorTest.parseColor("hsl(180, 12.5%, 87.5%)") is "rgb(219, 227, 227)"
PASS colorTest.parseColor("hsl(210, 12.5%, 87.5%)") is "rgb(219, 223, 227)"
PASS colorTest.parseColor("hsl(240, 12.5%, 87.5%)") is "rgb(219, 219, 227)"
PASS colorTest.parseColor("hsl(270, 12.5%, 87.5%)") is "rgb(223, 219, 227)"
PASS colorTest.parseColor("hsl(300, 12.5%, 87.5%)") is "rgb(227, 219, 227)"
PASS colorTest.parseColor("hsl(330, 12.5%, 87.5%)") is "rgb(227, 219, 223)"
PASS colorTest.parseColor("hsl(0, 25%, 87.5%)") is "rgb(231, 215, 215)"
PASS colorTest.parseColor("hsl(30, 25%, 87.5%)") is "rgb(231, 223, 215)"
PASS colorTest.parseColor("hsl(60, 25%, 87.5%)") is "rgb(231, 231, 215)"
PASS colorTest.parseColor("hsl(90, 25%, 87.5%)") is "rgb(223, 231, 215)"
PASS colorTest.parseColor("hsl(120, 25%, 87.5%)") is "rgb(215, 231, 215)"
PASS colorTest.parseColor("hsl(150, 25%, 87.5%)") is "rgb(215, 231, 223)"
PASS colorTest.parseColor("hsl(180, 25%, 87.5%)") is "rgb(215, 231, 231)"
PASS colorTest.parseColor("hsl(210, 25%, 87.5%)") is "rgb(215, 223, 231)"
PASS colorTest.parseColor("hsl(240, 25%, 87.5%)") is "rgb(215, 215, 231)"
PASS colorTest.parseColor("hsl(270, 25%, 87.5%)") is "rgb(223, 215, 231)"
PASS colorTest.parseColor("hsl(300, 25%, 87.5%)") is "rgb(231, 215, 231)"
PASS colorTest.parseColor("hsl(330, 25%, 87.5%)") is "rgb(231, 215, 223)"
PASS colorTest.parseColor("hsl(0, 37.5%, 87.5%)") is "rgb(235, 211, 211)"
PASS colorTest.parseColor("hsl(30, 37.5%, 87.5%)") is "rgb(235, 223, 211)"
PASS colorTest.parseColor("hsl(60, 37.5%, 87.5%)") is "rgb(235, 235, 211)"
PASS colorTest.parseColor("hsl(90, 37.5%, 87.5%)") is "rgb(223, 235, 211)"
PASS colorTest.parseColor("hsl(120, 37.5%, 87.5%)") is "rgb(211, 235, 211)"
PASS colorTest.parseColor("hsl(150, 37.5%, 87.5%)") is "rgb(211, 235, 223)"
PASS colorTest.parseColor("hsl(180, 37.5%, 87.5%)") is "rgb(211, 235, 235)"
PASS colorTest.parseColor("hsl(210, 37.5%, 87.5%)") is "rgb(211, 223, 235)"
PASS colorTest.parseColor("hsl(240, 37.5%, 87.5%)") is "rgb(211, 211, 235)"
PASS colorTest.parseColor("hsl(270, 37.5%, 87.5%)") is "rgb(223, 211, 235)"
PASS colorTest.parseColor("hsl(300, 37.5%, 87.5%)") is "rgb(235, 211, 235)"
PASS colorTest.parseColor("hsl(330, 37.5%, 87.5%)") is "rgb(235, 211, 223)"
PASS colorTest.parseColor("hsl(0, 50%, 87.5%)") is "rgb(239, 207, 207)"
PASS colorTest.parseColor("hsl(30, 50%, 87.5%)") is "rgb(239, 223, 207)"
PASS colorTest.parseColor("hsl(60, 50%, 87.5%)") is "rgb(239, 239, 207)"
PASS colorTest.parseColor("hsl(90, 50%, 87.5%)") is "rgb(223, 239, 207)"
PASS colorTest.parseColor("hsl(120, 50%, 87.5%)") is "rgb(207, 239, 207)"
PASS colorTest.parseColor("hsl(150, 50%, 87.5%)") is "rgb(207, 239, 223)"
PASS colorTest.parseColor("hsl(180, 50%, 87.5%)") is "rgb(207, 239, 239)"
PASS colorTest.parseColor("hsl(210, 50%, 87.5%)") is "rgb(207, 223, 239)"
PASS colorTest.parseColor("hsl(240, 50%, 87.5%)") is "rgb(207, 207, 239)"
PASS colorTest.parseColor("hsl(270, 50%, 87.5%)") is "rgb(223, 207, 239)"
PASS colorTest.parseColor("hsl(300, 50%, 87.5%)") is "rgb(239, 207, 239)"
PASS colorTest.parseColor("hsl(330, 50%, 87.5%)") is "rgb(239, 207, 223)"
PASS colorTest.parseColor("hsl(0, 62.5%, 87.5%)") is "rgb(243, 203, 203)"
PASS colorTest.parseColor("hsl(30, 62.5%, 87.5%)") is "rgb(243, 223, 203)"
PASS colorTest.parseColor("hsl(60, 62.5%, 87.5%)") is "rgb(243, 243, 203)"
PASS colorTest.parseColor("hsl(90, 62.5%, 87.5%)") is "rgb(223, 243, 203)"
PASS colorTest.parseColor("hsl(120, 62.5%, 87.5%)") is "rgb(203, 243, 203)"
PASS colorTest.parseColor("hsl(150, 62.5%, 87.5%)") is "rgb(203, 243, 223)"
PASS colorTest.parseColor("hsl(180, 62.5%, 87.5%)") is "rgb(203, 243, 243)"
PASS colorTest.parseColor("hsl(210, 62.5%, 87.5%)") is "rgb(203, 223, 243)"
PASS colorTest.parseColor("hsl(240, 62.5%, 87.5%)") is "rgb(203, 203, 243)"
PASS colorTest.parseColor("hsl(270, 62.5%, 87.5%)") is "rgb(223, 203, 243)"
PASS colorTest.parseColor("hsl(300, 62.5%, 87.5%)") is "rgb(243, 203, 243)"
PASS colorTest.parseColor("hsl(330, 62.5%, 87.5%)") is "rgb(243, 203, 223)"
PASS colorTest.parseColor("hsl(0, 75%, 87.5%)") is "rgb(247, 199, 199)"
PASS colorTest.parseColor("hsl(30, 75%, 87.5%)") is "rgb(247, 223, 199)"
PASS colorTest.parseColor("hsl(60, 75%, 87.5%)") is "rgb(247, 247, 199)"
PASS colorTest.parseColor("hsl(90, 75%, 87.5%)") is "rgb(223, 247, 199)"
PASS colorTest.parseColor("hsl(120, 75%, 87.5%)") is "rgb(199, 247, 199)"
PASS colorTest.parseColor("hsl(150, 75%, 87.5%)") is "rgb(199, 247, 223)"
PASS colorTest.parseColor("hsl(180, 75%, 87.5%)") is "rgb(199, 247, 247)"
PASS colorTest.parseColor("hsl(210, 75%, 87.5%)") is "rgb(199, 223, 247)"
PASS colorTest.parseColor("hsl(240, 75%, 87.5%)") is "rgb(199, 199, 247)"
PASS colorTest.parseColor("hsl(270, 75%, 87.5%)") is "rgb(223, 199, 247)"
PASS colorTest.parseColor("hsl(300, 75%, 87.5%)") is "rgb(247, 199, 247)"
PASS colorTest.parseColor("hsl(330, 75%, 87.5%)") is "rgb(247, 199, 223)"
PASS colorTest.parseColor("hsl(0, 87.5%, 87.5%)") is "rgb(251, 195, 195)"
PASS colorTest.parseColor("hsl(30, 87.5%, 87.5%)") is "rgb(251, 223, 195)"
PASS colorTest.parseColor("hsl(60, 87.5%, 87.5%)") is "rgb(251, 251, 195)"
PASS colorTest.parseColor("hsl(90, 87.5%, 87.5%)") is "rgb(223, 251, 195)"
PASS colorTest.parseColor("hsl(120, 87.5%, 87.5%)") is "rgb(195, 251, 195)"
PASS colorTest.parseColor("hsl(150, 87.5%, 87.5%)") is "rgb(195, 251, 223)"
PASS colorTest.parseColor("hsl(180, 87.5%, 87.5%)") is "rgb(195, 251, 251)"
PASS colorTest.parseColor("hsl(210, 87.5%, 87.5%)") is "rgb(195, 223, 251)"
PASS colorTest.parseColor("hsl(240, 87.5%, 87.5%)") is "rgb(195, 195, 251)"
PASS colorTest.parseColor("hsl(270, 87.5%, 87.5%)") is "rgb(223, 195, 251)"
PASS colorTest.parseColor("hsl(300, 87.5%, 87.5%)") is "rgb(251, 195, 251)"
PASS colorTest.parseColor("hsl(330, 87.5%, 87.5%)") is "rgb(251, 195, 223)"
PASS colorTest.parseColor("hsl(0, 100%, 87.5%)") is "rgb(255, 191, 191)"
PASS colorTest.parseColor("hsl(30, 100%, 87.5%)") is "rgb(255, 223, 191)"
PASS colorTest.parseColor("hsl(60, 100%, 87.5%)") is "rgb(255, 255, 191)"
PASS colorTest.parseColor("hsl(90, 100%, 87.5%)") is "rgb(223, 255, 191)"
PASS colorTest.parseColor("hsl(120, 100%, 87.5%)") is "rgb(191, 255, 191)"
PASS colorTest.parseColor("hsl(150, 100%, 87.5%)") is "rgb(191, 255, 223)"
PASS colorTest.parseColor("hsl(180, 100%, 87.5%)") is "rgb(191, 255, 255)"
PASS colorTest.parseColor("hsl(210, 100%, 87.5%)") is "rgb(191, 223, 255)"
PASS colorTest.parseColor("hsl(240, 100%, 87.5%)") is "rgb(191, 191, 255)"
PASS colorTest.parseColor("hsl(270, 100%, 87.5%)") is "rgb(223, 191, 255)"
PASS colorTest.parseColor("hsl(300, 100%, 87.5%)") is "rgb(255, 191, 255)"
PASS colorTest.parseColor("hsl(330, 100%, 87.5%)") is "rgb(255, 191, 223)"
PASS colorTest.parseColor("hsl(0, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 0%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 12.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 25%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 37.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 50%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 62.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 75%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 87.5%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(0, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(30, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(60, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(90, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(120, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(150, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(180, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(210, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(240, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(270, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(300, 100%, 100%)") is "rgb(255, 255, 255)"
PASS colorTest.parseColor("hsl(330, 100%, 100%)") is "rgb(255, 255, 255)"
PASS successfullyParsed is true

TEST COMPLETE
